% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{onek_genomes_tiny}
\alias{onek_genomes_tiny}
\title{1000 Genomes tiny subset}
\usage{
onek_genomes_tiny()
}
\value{
A character string to be passed to \code{\link[pcapred]{readinput}} as the file name.
}
\description{
The extdata directory contains three files wih the file root \code{1000G_tinysubset} that are read together with the \code{\link[pcapred]{readreference}} command.  This function gets their location on your system.
}
\keyword{datasets}
