% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{pcapred.ref-package}
\alias{pcapred.ref}
\alias{pcapred.ref-package}
\title{pcapred.ref: Data for pcapred}
\description{
Use with package "pcapred", available at https://github.com/danjlawson/pcapred. Provides 18 PCs from the UK Biobank as reference data, as well as a small subset from 1000 Genomes for use as a bim/bed/fam plink format example dataset.
}
\details{
This is the essential reference data for "pcapred", written by Daniel Lawson (dan.lawson@bristol.ac.uk) and supporting a paper by Aliya Sarmanova, Tim Morris and Daniel Lawson (to appear).

See \url{https://github.com/danjlawson/pcapred} for details.
}
\examples{
\dontrun{
library("pcapred")
ref=readreference(pcapred.ref::ukb_pcs_18())

}


}
\references{
To follow
}
\seealso{
\code{\link[pcapred]{readreference}}
}
\author{
\strong{Maintainer}: Daniel Lawson \email{dan.lawson@bristol.ac.uk}

}
\keyword{internal}
